add generics
